move config map watcher to k8scfgwatch
